Consolidate Metadata namespace into Routing namespace
Merged `Nalix.Network.Routing.Metadata` into `Nalix.Network.Routing`
to simplify the namespace structure and consolidate related
classes. Updated all references to the old namespace in
`PacketHandlerCompiler.cs`, `PacketHandler.cs`, `PacketContext.cs`,
`PacketMetadata.cs`, and `PacketMetadataBuilder.cs`.

Removed unnecessary `using Nalix.Network.Routing.Metadata;`
directives from affected files. This change improves code
maintainability by reducing the number of namespaces and making
it easier to manage related classes.
